server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Mateusz Jakubowski <mateusz.jakubowski@swmansion.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg = "^3.1.14"
python-decouple = "^3.8"
strawberry-graphql-django = "^0.28.1"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
django-extensions = "^3.2.3"
stripe = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
